track_eyes_face/requirements.txt
streamlit
matplotlib==3.9.4
mediapipe==0.10.21
opencv-python-headless==4.11.0
numpy==1.26.4
Pillow==11.3.0
